pyproject.toml
[project]
name = "travelguide"
version = "0.0.0.1"
description = "travelguide"
readme = "README.md"
authors = [
  { name = "Carlos Vinícius Teixeira de Souza", email = "carlosvtsdev@gmail.com" },
]
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "distro==1.9.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.10.0",
    "openai==1.86.0",
    "pydantic==2.11.7",
    "pydantic_core==2.33.2",
    "python-dotenv==1.1.0",
    "sniffio==1.3.1",
    "tqdm==4.67.1",
    "typing-inspection==0.4.1",
    "typing_extensions==4.14.0",
    "urllib3==2.4.0",
]

[project.urls]
website = "https://example.com"

[project.scripts]
travelguide = "travelguide.cli:run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

